refactor: split quiz orchestrator into modular structure
- Split large orchestrator.py (1,214 lines) into focused modules
- Created /orchestrator folder with 4 specialized files:
  - core.py: Background task safety and timeout utilities
  - content_extraction.py: Canvas/manual content processing workflow
  - question_generation.py: Module-based generation with batch tracking
  - export.py: All-or-nothing Canvas export with rollback
- Maintained backward compatibility through __init__.py exports
- All 508 tests pass, no breaking changes to existing API
- Improved maintainability and separation of concerns
