+++ backend/src/quiz/orchestrator/__init__.py
@@ -1,0 +1,23 @@
+"""
+Quiz orchestration module providing unified workflow management.
+
+This module exposes the main orchestration functions for content extraction,
+question generation, and Canvas export workflows while maintaining a clean
+separation of concerns through modular architecture.
+"""
+
+# Export core background task safety utilities
+# Export main orchestration workflows
+from .content_extraction import orchestrate_content_extraction
+from .core import safe_background_orchestration
+from .export import orchestrate_quiz_export_to_canvas
+from .question_generation import orchestrate_quiz_question_generation
+
+__all__ = [
+    # Main orchestration functions (public API)
+    "orchestrate_content_extraction",
+    "orchestrate_quiz_question_generation",
+    "orchestrate_quiz_export_to_canvas",
+    # Background task utilities
+    "safe_background_orchestration",
+]
